pyproject.toml
[tool.poetry]
name = "codingcontest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = ">=1.23.4"
pandas = ">=1.5.1"
tqdm = ">=4.64.1"
requests = ">=2.28.1"
loguru = ">=0.6.0"
python-dotenv = ">=0.21.0"
typer = ">=0.7.0"
gitpython = ">=3.1.29"
types-requests = "^2.28.11.16"

[tool.poetry.group.dev.dependencies]
black = ">=22.10.0"
jupyterlab = ">=3.5.0"
matplotlib = ">=3.6.2"
mypy = ">=0.982"
pdbpp = ">=0.10.3"
scikit-learn = ">=1.1.3"
seaborn = ">=0.12.1"
Flake8-pyproject = ">=1.1.0.post0"
types-requests = ">=2.28.11.2"
ruff = ">=0.0.259"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
next-level = 'codingcontest.next_level:run'
submit-solutions = 'codingcontest.submit_solutions:run'

[tool.black]
line-length = 99

[tool.mypy]
files = ["codingcontest"]
python_version = "3.10"
warn_unreachable = true
warn_return_any = true
disallow_untyped_defs = true
show_error_codes = true
ignore_missing_imports = false
exclude = [".git/", ".venv/", "__pycache__"]
follow_imports = "normal"
pretty = true
warn_incomplete_stub = true

[tool.ruff]
line-length = 99
target-version = "py311"

select = ["ALL"]

ignore = [
    "ANN101", # dont require self typing
    "ANN102", # dont require cls typing
    "COM812", # dont add trailing commas always
    "D100",   # dont force docstring presence
    "D101",   # dont force docstring presence
    "D102",   # dont force docstring presence
    "D103",   # dont force docstring presence
    "D104",   # dont force docstring presence
    "D105",   # dont force docstring presence
    "D106",   # dont force docstring presence
    "D203",   # conflict with other rule, this silences the warning
    "D213",   # conflict with other rule, this silences the warning
    "ERA",    # commenting out code is ok
    "G004",   # allow f-string in logging
    "S101",   # allow asserts
    "TRY400", # logger.error is not critical with loguru
]

unfixable = [
    "ARG",  # Don't fix any "unused arguments" errors
    "B007", # Don't automatically prefix loop control variable with "_" if unused
    "ERA",  # Don't delete commented code
    "F401", # Don't remove unused imports
    "F504", # Don't remove unused % format arguments
    "F522", # Don't remove unused .format arguments
    "F841", # Don't remove unused variables automatically
]

[tool.ruff.per-file-ignores]


[tool.ruff.isort]
known-first-party = ["codingcontest"]
